feat: externalize function
