+++ src/app/simulation/etapes/WizardForm.tsx
@@ -2,9 +2,10 @@ import { Base64 } from "js-base64";
 import { useRouter } from "next/navigation";
 import { useEffect, useState } from "react";
 import { useWizard } from "react-use-wizard";
-import { z, type ZodFormattedError } from "zod";
+import { type z, type ZodFormattedError } from "zod";
 
 import { usePacoupaSessionStorage } from "@/lib/client/usePacoupaSessionStorage";
+import { createObjectFromFormData } from "@/lib/form-data";
 
 import { type SimulationSchema } from "../schema";
 import { ButtonsFunnel } from "./ButtonsFunnel";
@@ -38,21 +39,6 @@ type Props = {
   schema: z.AnyZodObject;
 };
 
-const createJSONFromFormData = (data: FormData, zodSchema: z.AnyZodObject) => {
-  // eslint-disable-next-line @typescript-eslint/no-unsafe-argument
-  const keys = Object.keys(zodSchema.shape) as unknown as Array<keyof z.infer<typeof zodSchema>>;
-  // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
-  const arrayKeys = keys.filter(key => zodSchema.shape[key] instanceof z.ZodArray);
-
-  const values = Object.fromEntries(
-    Array.from(data.keys()).map(key => [key, arrayKeys.includes(key) ? data.getAll(key) : data.get(key)]),
-  );
-
-  console.debug("Debug values createJSONFromFormData", JSON.stringify(values));
-
-  return values;
-};
-
 export const WizardForm = ({ schema, render }: Props) => {
   const { nextStep, isLastStep } = useWizard();
   const [errors, setErrors] = useState<ZodFormattedError<{ [x: string]: unknown }, string>>();
@@ -73,7 +59,7 @@ export const WizardForm = ({ schema, render }: Props) => {
     event.preventDefault();
 
     const formData = new FormData(event.currentTarget);
-    const validation = schema.safeParse(createJSONFromFormData(formData, schema));
+    const validation = schema.safeParse(createObjectFromFormData(schema)(formData));
 
     if (validation.success) {
       const nextStore = { ...store, ...validation.data } as SimulationSchema;
